+++ pkg/package_manager.go
@@ -17,6 +17,7 @@ import (
 	"github.com/wttech/aemc/pkg/common/stringsx"
 	"github.com/wttech/aemc/pkg/common/timex"
 	"github.com/wttech/aemc/pkg/common/tplx"
+	"github.com/wttech/aemc/pkg/content"
 	"github.com/wttech/aemc/pkg/pkg"
 	"io/fs"
 	"os"
@@ -174,8 +175,8 @@ func (pm *PackageManager) Create(pid string, rootPaths []string, filterFile stri
 			}).
 			Post(ExecPath + "?cmd=create")
 	} else {
-		tmpDir := pathx.RandomTemporaryPathName(pm.config.TmpDir, "tmppck")
-		tmpFile := pathx.RandomTemporaryFileName(pm.config.TmpDir, "tmppck", ".zip")
+		tmpDir := pathx.RandomTemporaryPathName(pm.tmpDir(), "tmppck")
+		tmpFile := pathx.RandomTemporaryFileName(pm.tmpDir(), "tmppck", ".zip")
 		defer func() {
 			_ = pathx.DeleteIfExists(tmpDir)
 			_ = pathx.DeleteIfExists(tmpFile)
@@ -566,6 +567,108 @@ func (pm *PackageManager) Delete(remotePath string) error {
 	return nil
 }
 
+func (pm *PackageManager) tmpDir() string {
+	return pm.instance.manager.aem.baseOpts.TmpDir
+}
+
+func (pm *PackageManager) DownloadPackage(pid string, roots []string, filter string) (string, error) {
+	if pid == "" {
+		pid = "my_packages:aemc_content:" + time.Now().Format("2006.102.304") + "-SNAPSHOT"
+	}
+	remotePath, err := pm.Create(pid, roots, filter)
+	if err != nil {
+		return "", err
+	}
+	defer func() {
+		_ = pm.Delete(remotePath)
+	}()
+	if err := pm.Build(remotePath); err != nil {
+		return "", err
+	}
+	tmpResultFile := filepath.Join(pm.instance.manager.aem.baseOpts.TmpDir, filepath.Base(remotePath))
+	if err := pm.Download(remotePath, tmpResultFile); err != nil {
+		return "", err
+	}
+	return tmpResultFile, nil
+}
+
+func (pm *PackageManager) DownloadContent(pid string, root string, roots []string, filter string, clean bool, unpack bool) error {
+	tmpResultFile, err := pm.DownloadPackage(pid, roots, filter)
+	if err != nil {
+		return err
+	}
+	tmpResultDir := pathx.RandomTemporaryPathName(pm.tmpDir(), "pkg_download_content")
+	defer func() {
+		_ = pathx.DeleteIfExists(tmpResultDir)
+		_ = pathx.DeleteIfExists(tmpResultFile)
+	}()
+	if unpack {
+		if err = filex.Unarchive(tmpResultFile, tmpResultDir); err != nil {
+			return err
+		}
+		if err := pathx.Ensure(root); err != nil {
+			return err
+		}
+		before, _, _ := strings.Cut(root, content.JCRRoot)
+		contentManager := pm.instance.manager.aem.contentManager
+		if clean {
+			if err = contentManager.BeforeClean(root); err != nil {
+				return err
+			}
+		}
+		if err = filex.CopyDir(filepath.Join(tmpResultDir, content.JCRRoot), before+content.JCRRoot); err != nil {
+			return err
+		}
+		if clean {
+			if err = contentManager.Clean(root); err != nil {
+				return err
+			}
+		}
+	} else {
+		if err = filex.Copy(tmpResultFile, filepath.Join(root, filepath.Base(tmpResultFile)), true); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (pm *PackageManager) CopyContent(destInstance *Instance, pid string, roots []string, filter string, clean bool) error {
+
+	var tmpResultFile string
+	if clean {
+		tmpResultFile = pathx.RandomTemporaryFileName(pm.tmpDir(), "pkg_copy_content", ".zip")
+		tmpResultDir := pathx.RandomTemporaryPathName(pm.tmpDir(), "pkg_copy_content")
+		defer func() {
+			_ = pathx.DeleteIfExists(tmpResultDir)
+			_ = pathx.DeleteIfExists(tmpResultFile)
+		}()
+		if err := pm.DownloadContent(filepath.Join(tmpResultDir, content.JCRRoot), "", roots, filter, true, true); err != nil {
+			return err
+		}
+		if err := filex.Archive(tmpResultDir, tmpResultFile); err != nil {
+			return err
+		}
+	} else {
+		var err error
+		tmpResultFile, err = pm.DownloadPackage(pid, roots, filter)
+		if err != nil {
+			return err
+		}
+	}
+	defer func() { _ = pathx.DeleteIfExists(tmpResultFile) }()
+	remotePath, err := destInstance.PackageManager().Upload(tmpResultFile)
+	if err != nil {
+		return err
+	}
+	defer func() {
+		_ = destInstance.PackageManager().Delete(remotePath)
+	}()
+	if err = destInstance.PackageManager().Install(remotePath); err != nil {
+		return err
+	}
+	return nil
+}
+
 const (
 	MgrPath         = "/crx/packmgr"
 	ServicePath     = MgrPath + "/service"
@@ -575,4 +678,6 @@ const (
 	IndexPath       = MgrPath + "/index.jsp"
 	ExecPath        = ServicePath + "/exec.json"
 	UpdatePath      = MgrPath + "/update.jsp"
+
+	FilterXML = "filter.xml"
 )
